Beginnings of a PETSCII telnet game server
There is a LOT of work to do yet, but this is a starting point at least for
a general purpose game server along with a PETSCII telnet server that can
interface with it. The core Games::Catan code will need to change to better
facilitate players adhering to the rules as well as letting players know of game
events as they happen.